// src/main/resources/site/content-types/office-data/office-data.ts
export interface OfficeData {
  /**
   * Checksum
   */
  checksum?: string;

  /**
   * Skriv inn ønsket kort-url
   */
  customPath?: string;

  /**
   * Enhet
   */
  enhet: {
    /**
     * EnhetId
     */
    enhetId: number;

    /**
     * Navn
     */
    navn: string;

    /**
     * EnhetNr
     */
    enhetNr: string;

    /**
     * Antall Ressurser
     */
    antallRessurser?: number;

    /**
     * Status
     */
    status?: string;

    /**
     * Org Nivå
     */
    orgNivaa?: string;

    /**
     * Type
     */
    type: string;

    /**
     * Organisasjonsnummer
     */
    organisasjonsnummer?: string;

    /**
     * Under Etablering Dato
     */
    underEtableringDato?: string;

    /**
     * Aktiveringsdato
     */
    aktiveringsdato?: string;

    /**
     * Under Avvikling Dato
     */
    underAvviklingDato?: string;

    /**
     * Nedleggelsesdato
     */
    nedleggelsesdato?: string;

    /**
     * Oppgavebehandler
     */
    oppgavebehandler?: string;

    /**
     * Versjon
     */
    versjon?: number;

    /**
     * Sosiale Tjenester
     */
    sosialeTjenester?: string;

    /**
     * Kanalstrategi
     */
    kanalstrategi?: string;

    /**
     * Org Nr Til Kommunalt Nav Kontor
     */
    orgNrTilKommunaltNavKontor?: string;
  };

  /**
   * Overordnet Enhet
   */
  overordnetEnhet?: string;

  /**
   * Habilitetskontor
   */
  habilitetskontor?: Array<string>;

  /**
   * Kontaktinformasjon
   */
  kontaktinformasjon?: {
    /**
     * Id
     */
    id?: number;

    /**
     * EnhetNr
     */
    enhetNr?: string;

    /**
     * Telefonnummer
     */
    telefonnummer?: string;

    /**
     * Telefonnummer Kommentar
     */
    telefonnummerKommentar?: string;

    /**
     * Faksnummer
     */
    faksnummer?: string;

    /**
     * E-post
     */
    epost?: {
      /**
       * Adresse
       */
      adresse?: string;

      /**
       * Kommentar
       */
      kommentar?: string;

      /**
       * kunIntern
       */
      kunIntern?: string;
    };

    /**
     * Postadresse
     */
    postadresse?: {
      /**
       * Type
       */
      type?: string;

      /**
       * Postnummer
       */
      postnummer?: string;

      /**
       * Poststed
       */
      poststed?: string;

      /**
       * Gatenavn
       */
      gatenavn?: string;

      /**
       * Husnummer
       */
      husnummer?: string;

      /**
       * Husbokstav
       */
      husbokstav?: string;

      /**
       * Adresse Tilleggsnavn
       */
      adresseTilleggsnavn?: string;

      /**
       * Postboksnummer
       */
      postboksnummer?: string;

      /**
       * Postboksanlegg
       */
      postboksanlegg?: string;
    };

    /**
     * Besøksadresse
     */
    besoeksadresse?: {
      /**
       * Type
       */
      type?: string;

      /**
       * Postnummer
       */
      postnummer?: string;

      /**
       * Poststed
       */
      poststed?: string;

      /**
       * Gatenavn
       */
      gatenavn?: string;

      /**
       * Husnummer
       */
      husnummer?: string;

      /**
       * Husbokstav
       */
      husbokstav?: string;

      /**
       * Adresse Tilleggsnavn
       */
      adresseTilleggsnavn?: string;
    };

    /**
     * Spesielle Opplysninger
     */
    spesielleOpplysninger?: string;

    /**
     * Publikumsmottak
     */
    publikumsmottak?: Array<{
      /**
       * Id
       */
      id?: number;

      /**
       * Stedsbeskrivelse
       */
      stedsbeskrivelse?: string;

      /**
       * Besøksadresse
       */
      besoeksadresse?: {
        /**
         * Type
         */
        type?: string;

        /**
         * Postnummer
         */
        postnummer?: string;

        /**
         * Poststed
         */
        poststed?: string;

        /**
         * Gatenavn
         */
        gatenavn?: string;

        /**
         * Husnummer
         */
        husnummer?: string;

        /**
         * Husbokstav
         */
        husbokstav?: string;

        /**
         * Adresse Tilleggsnavn
         */
        adresseTilleggsnavn?: string;
      };

      /**
       * Åpningstider
       */
      aapningstider?: Array<{
        /**
         * Id
         */
        id?: number;

        /**
         * Dag
         */
        dag?: string;

        /**
         * Dato
         */
        dato?: string;

        /**
         * Fra
         */
        fra?: string;

        /**
         * Til
         */
        til?: string;

        /**
         * Kommentar
         */
        kommentar?: string;

        /**
         * Stengt
         */
        stengt?: string;
      }>;
    }>;
  };

  /**
   * Fant du det du lette etter?
   */
  feedbackToggle: boolean;

  /**
   * Chatbot Frida
   */
  chatbotToggle: boolean;
}